advreport: consolidated cancel and generate buttons under one class
